fastapi-prefect/api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
